some successful refactoring
