+++ app/views/filter_home/index.html.erb
@@ -1,13 +1,5 @@
 
-<nav>
-  <div>
-    <%= image_tag("logo-2019-inverted.png", :alt => "logos", class: "logo") %>
 
-    <h2>add an icon for assistance<h2>
-    <h2>add search functionality here</h2>
-  </div>
-</nav>
-
-<main>
+<%# <main> %>
   <%= react_component("FilterHome", props: {siteNames: @site_names})  %>
-</main>
+<%# </main> %>
