+++ public/packs/manifest.json
@@ -1,8 +1,8 @@
 {
-  "application.js": "/packs/application-bf8292a7b02f1c3ba662.js",
-  "application.js.map": "/packs/application-bf8292a7b02f1c3ba662.js.map",
-  "filter-home-bundle.js": "/packs/filter-home-bundle-bf8292a7b02f1c3ba662.js",
-  "filter-home-bundle.js.map": "/packs/filter-home-bundle-bf8292a7b02f1c3ba662.js.map",
-  "hello_react.js": "/packs/hello_react-bf8292a7b02f1c3ba662.js",
-  "hello_react.js.map": "/packs/hello_react-bf8292a7b02f1c3ba662.js.map"
+  "application.js": "/packs/application-804e5b2092d44e3be6bb.js",
+  "application.js.map": "/packs/application-804e5b2092d44e3be6bb.js.map",
+  "filter-home-bundle.js": "/packs/filter-home-bundle-804e5b2092d44e3be6bb.js",
+  "filter-home-bundle.js.map": "/packs/filter-home-bundle-804e5b2092d44e3be6bb.js.map",
+  "hello_react.js": "/packs/hello_react-804e5b2092d44e3be6bb.js",
+  "hello_react.js.map": "/packs/hello_react-804e5b2092d44e3be6bb.js.map"
 }
